Updated song for fixed pattern steps
Song has been updated to the new fixed step format and is
built from sub-patterns of 8 rows, and 4 of those for a
pattern (ala tracker like). Labels have been provided to
note which sub-patterns and patterns repeat.

Unfortunately, the weird timings for the Arduino Uno make
it hard to test since the glides are all off and the tempo
isn't consistent. Since I removed the fast part I may also
add in another melody, depending on how much space is available
while also trying to integrate it into the original Intro song.